models: move game specific artifact models to their own file
